+++ changelog.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add documentation (`changelog.md` and `.markdownlint.json`).
 - Configure Path aliases [DominiCode](https://www.youtube.com/watch?v=f5aEPnoh9rs).
+- Create Receipt Reports.
+- Create Lessons Reports.
 
 ## [1.2.0] - 2020-04-24